Lgm_Utils: updated int->long to allow for longer arrays

+++ libLanlGeoMag/Lgm/Lgm_Utils.h
@@ -1,9 +1,9 @@
 #ifndef LGM_UTILS_H
 #define LGM_UTILS_H
 
-void    Lgm_LogSpace( double start, double stop, int num, double *array );
-void    Lgm_LinSpace(double start, double stop, int num, double *array);
-int     Lgm_Bisect(double *data, double value, int len);
+void    Lgm_LogSpace( double start, double stop, long num, double *array );
+void    Lgm_LinSpace(double start, double stop, long num, double *array);
+long     Lgm_Bisect(double *data, double value, long len);
